FIX: Move cache_ctrl to CachingNode
