+++ podpac/core/node.py
@@ -31,7 +31,6 @@ from podpac.core.utils import NodeTrait
 from podpac.core.utils import hash_alg
 from podpac.core.coordinates import Coordinates
 from podpac.core.style import Style
-from podpac.core.cache import CacheCtrl, get_default_cache_ctrl, make_cache_ctrl, S3CacheStore, DiskCacheStore
 from podpac.core.managers.multi_threading import thread_manager
 
 _logger = logging.getLogger(__name__)
@@ -107,8 +106,6 @@ class Node(tl.HasTraits):
     force_eval: bool
         Default is False. Should the node's cached output be updated from the source data? If True it ignores the cache
         when computing outputs but puts results into the cache (thereby updating the cache)
-    cache_ctrl: :class:`podpac.core.cache.cache.CacheCtrl`
-        Class that controls caching. If not provided, uses default based on settings.
     dtype : type
         The numpy datatype of the output. Currently only ``float`` is supported.
     style : :class:`podpac.Style`
@@ -138,7 +135,6 @@ class Node(tl.HasTraits):
     dtype = tl.Enum([float], default_value=float)
     cache_output = tl.Bool()
     force_eval = tl.Bool(False)
-    cache_ctrl = tl.Instance(CacheCtrl, allow_none=True)
 
     # list of attribute names, used by __repr__ and __str__ to display minimal info about the node
     # e.g. data sources use ['source']
@@ -170,14 +166,10 @@ class Node(tl.HasTraits):
     def _cache_output_default(self):
         return settings["CACHE_NODE_OUTPUT_DEFAULT"]
 
-    @tl.default("cache_ctrl")
-    def _cache_ctrl_default(self):
-        return get_default_cache_ctrl()
 
     # debugging
     _requested_coordinates = tl.Instance(Coordinates, allow_none=True)
     _output = tl.Instance(UnitsDataArray, allow_none=True)
-    _from_cache = tl.Bool(allow_none=True, default_value=None)
     # Flag that is True if the Node was run multi-threaded, or None if the question doesn't apply
     _multi_threaded = tl.Bool(allow_none=True, default_value=None)
 
@@ -187,10 +179,6 @@ class Node(tl.HasTraits):
 
     def __init__(self, **kwargs):
         """Do not overwrite me"""
-
-        # Shortcut for users to make setting the cache_ctrl simpler:
-        if "cache_ctrl" in kwargs and isinstance(kwargs["cache_ctrl"], list):
-            kwargs["cache_ctrl"] = make_cache_ctrl(kwargs["cache_ctrl"])
 
         tkwargs = self._first_init(**kwargs)
 
@@ -672,137 +660,7 @@ class Node(tl.HasTraits):
             return True
         return self.hash != other.hash
 
-    # -----------------------------------------------------------------------------------------------------------------
-    # Caching Interface
-    # -----------------------------------------------------------------------------------------------------------------
-
-    def get_cache(self, key, coordinates=None):
-        """
-        Get cached data for this node.
-
-        Parameters
-        ----------
-        key : str
-            Key for the cached data, e.g. 'output'
-        coordinates : podpac.Coordinates, optional
-            Coordinates for which the cached data should be retrieved. Omit for coordinate-independent data.
-
-        Returns
-        -------
-        data : any
-            The cached data.
-
-        Raises
-        ------
-        NodeException
-            Cached data not found.
-        """
-
-        try:
-            self.definition
-        except NodeDefinitionError as e:
-            raise NodeException("Cache unavailable, %s (key='%s')" % (e.args[0], key))
-
-        if self.cache_ctrl is None or not self.has_cache(key, coordinates=coordinates):
-            raise NodeException("cached data not found for key '%s' and coordinates %s" % (key, coordinates))
-
-        return self.cache_ctrl.get(self, key, coordinates=coordinates)
-
-    def put_cache(self, data, key, coordinates=None, expires=None, overwrite=True):
-        """
-        Cache data for this node.
-
-        Parameters
-        ----------
-        data : any
-            The data to cache.
-        key : str
-            Unique key for the data, e.g. 'output'
-        coordinates : podpac.Coordinates, optional
-            Coordinates that the cached data depends on. Omit for coordinate-independent data.
-        expires : float, datetime, timedelta
-            Expiration date. If a timedelta is supplied, the expiration date will be calculated from the current time.
-        overwrite : bool, optional
-            Overwrite existing data, default True.
-
-        Raises
-        ------
-        NodeException
-            Cached data already exists (and overwrite is False)
-        """
-
-        try:
-            self.definition
-        except NodeDefinitionError as e:
-            raise NodeException("Cache unavailable, %s (key='%s')" % (e.args[0], key))
-
-        if self.cache_ctrl is None:
-            return
-
-        if not overwrite and self.has_cache(key, coordinates=coordinates):
-            raise NodeException("Cached data already exists for key '%s' and coordinates %s" % (key, coordinates))
-
-        with thread_manager.cache_lock:
-            self.cache_ctrl.put(self, data, key, coordinates=coordinates, expires=expires, update=overwrite)
-
-    def has_cache(self, key, coordinates=None):
-        """
-        Check for cached data for this node.
-
-        Parameters
-        ----------
-        key : str
-            Key for the cached data, e.g. 'output'
-        coordinates : podpac.Coordinates, optional
-            Coordinates for which the cached data should be retrieved. Omit for coordinate-independent data.
-
-        Returns
-        -------
-        bool
-            True if there is cached data for this node, key, and coordinates.
-        """
-
-        try:
-            self.definition
-        except NodeDefinitionError as e:
-            raise NodeException("Cache unavailable, %s (key='%s')" % (e.args[0], key))
-
-        if self.cache_ctrl is None:
-            return False
-
-        with thread_manager.cache_lock:
-            return self.cache_ctrl.has(self, key, coordinates=coordinates)
-
-    def rem_cache(self, key, coordinates=None, mode="all"):
-        """
-        Clear cached data for this node.
-
-        Parameters
-        ----------
-        key : str
-            Delete cached objects with this key. If `'*'`, cached data is deleted for all keys.
-        coordinates : podpac.Coordinates, str, optional
-            Default is None. Delete cached objects for these coordinates. If `'*'`, cached data is deleted for all
-            coordinates, including coordinate-independent data. If None, will only affect coordinate-independent data.
-        mode: str, optional
-            Specify which cache stores are affected. Default 'all'.
-
-
-        See Also
-        ---------
-        `podpac.core.cache.cache.CacheCtrl.clear` to remove ALL cache for ALL nodes.
-        """
-
-        try:
-            self.definition
-        except NodeDefinitionError as e:
-            raise NodeException("Cache unavailable, %s (key='%s')" % (e.args[0], key))
-
-        if self.cache_ctrl is None:
-            return
-
-        self.cache_ctrl.rem(self, item=key, coordinates=coordinates, mode=mode)
-
+    
     # --------------------------------------------------------#
     #  Class Methods (Deserialization)
     # --------------------------------------------------------#
@@ -1338,34 +1196,7 @@ def _process_kwargs(name, d, definition, nodes):
     nodes[name] = node_class(**kwargs)
 
 
-# --------------------------------------------------------#
-#  Mixins
-# --------------------------------------------------------#
 
-
-class NoCacheMixin(tl.HasTraits):
-    """Mixin to use no cache by default."""
-
-    cache_ctrl = tl.Instance(CacheCtrl, allow_none=True)
-
-    @tl.default("cache_ctrl")
-    def _cache_ctrl_default(self):
-        return CacheCtrl([])
-
-
-class DiskCacheMixin(tl.HasTraits):
-    """Mixin to add disk caching to the Node by default."""
-
-    cache_ctrl = tl.Instance(CacheCtrl, allow_none=True)
-
-    @tl.default("cache_ctrl")
-    def _cache_ctrl_default(self):
-        # get the default cache_ctrl and addd a disk cache store if necessary
-        default_ctrl = get_default_cache_ctrl()
-        stores = default_ctrl._cache_stores
-        if not any(isinstance(store, DiskCacheStore) for store in default_ctrl._cache_stores):
-            stores.append(DiskCacheStore())
-        return CacheCtrl(stores)
 
 
 # --------------------------------------------------------#
